Also run tests with Immix
Now both MarkSweep and Immix are supported.  We can run tests with both
plans.  The focus of `test-mmtk.rb` is command line options and
MMTk-specific run-time features.  We can use the default plan (which is
currently MarkSweep, but is subject to change) instead of hard-coding
MarkSweep.

Some of the test-all tests still fail.  Nevertheless, we can enable
Immix for it, too, to reveal more problems.